writing reservations.js in routes

+++ src/backend/api/routes/reservations.js
@@ -28,7 +28,14 @@ router.post("/add", async(req, res) => {
     try {
         if(!body.name) throw new CustomError(Enum.HTTP_CODES.BAD_REQUEST, "Validation Error!", "name fields must be fiiled");
         let reservations = new Reservations({
-            name:body.name,
+            tenant_id:body.tenant_id,
+            field_id:body.field_id,
+            customer_id:body.customer_id,
+            start:body.start,
+            end:body.end,
+            price:body.price,
+            status:body.status,
+            client_request_id:body.client_request_id,
             is_active:true,
             created_by: req.user?.id
         });
@@ -51,13 +58,22 @@ router.post("/update", async (req, res) => {
   let body = req.body;
   try {
     if (!body._id) throw new CustomError(Enum.HTTP_CODES.BAD_REQUEST, "Validation Error!", "_id field must be filled");
+    if (!field._id) throw new CustomError(Enum.HTTP_CODES.BAD_REQUEST, "Validation Error!", "field._id field must be filled");
+    if (!customer._id) throw new CustomError(Enum.HTTP_CODES.BAD_REQUEST, "Validation Error!", "customer._id field must be filled");
 
     let updates = {};
 
-    if (body.name) updates.name = body.name;
+    if (body.tenant_id) updates.tenant_id = body.tenant_id;
+    if (body.field_id) updates.field_id = body.field_id;
+    if (body.customer_id) updates.customer_id = body.customer_id;
+    if (body.start) updates.start = body.start;
+    if (body.end) updates.end = body.end;
+    if (body.price) updates.price = body.price;
+    if (body.status) updates.status = body.status;
+    if (body.client_request_id) updates.client_request_id = body.client_request_id;
     if (typeof body.is_active === "boolean") updates.is_active = body.is_active;
 
-    await Reservations.updateOne({ _id: body._id }, updates);
+    await Reservations.updateOne({ _id: body._id, tenant_id:body.tenant_id, field_id:body.field_id, customer_id:body.customer_id}, updates);
 
     AuditLogs.info(req.user?.email, "Reservations", "Update", { _id: body._id, ...updates });
 
@@ -74,8 +90,10 @@ router.post("/delete", async (req, res) => {
   let body = req.body;
   try {
     if (!body._id) throw new CustomError(Enum.HTTP_CODES.BAD_REQUEST,"Validation Error!", "_id field must be filled");
+    if (!field._id) throw new CustomError(Enum.HTTP_CODES.BAD_REQUEST, "Validation Error!", "field._id field must be filled");
+    if (!customer._id) throw new CustomError(Enum.HTTP_CODES.BAD_REQUEST, "Validation Error!", "customer._id field must be filled");
 
-    await Reservations.deleteOne({ _id: body._id }); 
+    await Reservations.deleteOne({ _id: body._id, tenant_id:body.tenant_id, field_id:body.field_id, customer_id:body.customer_id }); 
 
     AuditLogs.info(req.user?.email, "Reservations", "Delete", { _id: body._id });
